Switched to size_t over int for size related returns

+++ Engine/Core/Animation.cpp
@@ -19,7 +19,7 @@ void Animation::AddKeyFrame(KeyFrame Keyframe) {
 	keyframes.push_back(Keyframe);
 }
 
-int Animation::GetKeyFrameSize() {
+size_t Animation::GetKeyFrameSize() {
 	return keyframes.size();
 }
