+++ Engine/Core/Scene/SceneManager.h
@@ -7,7 +7,7 @@
 namespace SceneManager
 {
 	void Init();
-	int CreateScene(Scene scene);
+	size_t CreateScene(Scene scene);
 	void LoadScene(int sceneNumber);
 	void Update(float dt);
 	void Render();
